m3d_color.hh: fix compilation with MSVC

+++ m3d_color.hh
@@ -64,13 +64,13 @@ public:
 		m3d_color temp(*this);
 		tempc = temp.mycolor.channels[R_CHANNEL] * other.mycolor.channels[R_CHANNEL];
 		tempc >>= 8;
-		temp.mycolor.channels[R_CHANNEL] = tempc;
+		temp.mycolor.channels[R_CHANNEL] = (uint8_t)tempc;
 		tempc = temp.mycolor.channels[G_CHANNEL] * other.mycolor.channels[G_CHANNEL];
 		tempc >>= 8;
-		temp.mycolor.channels[G_CHANNEL] = tempc;
+		temp.mycolor.channels[G_CHANNEL] = (uint8_t)tempc;
 		tempc = temp.mycolor.channels[B_CHANNEL] * other.mycolor.channels[B_CHANNEL];
 		tempc >>= 8;
-		temp.mycolor.channels[B_CHANNEL] = tempc;
+		temp.mycolor.channels[B_CHANNEL] = (uint8_t)tempc;
 		return temp;
 	}
